conn: fix two size_t/int truncation issues found while scoping #4940
Both were flagged in #4940's investigation as small, independent
fixes worth doing regardless of the larger (and not worth pursuing)
Connection::write() interface question.

gnutls.c (tls_socket_write): same size_t-into-int narrowing pattern
as Coverity CID 392753/392751 (raw.c/tunnel.c), though not currently
flagged itself. Explicit cast documents the narrowing is safe, given
the sole caller already bounds count to INT_MAX.

zstrm.c (zstrm_write): distinct bug, not a Coverity CID. The old
`(rc <= 0) ? 1 : rc` clamp silently reported a false "1 byte written"
if count ever overflowed INT_MAX, instead of surfacing the failure.
Replaced with an explicit overflow check returning -1.

AI disclosure: implemented with Claude Code's assistance, following
up on the investigation already posted to #4940 on 2026-07-21.

+++ conn/gnutls.c
@@ -1127,7 +1127,8 @@ static int tls_socket_write(struct Connection *conn, const char *buf, size_t cou
     sent += rc;
   } while (sent < count);
 
-  return sent;
+  // Narrowed to int for the Connection::write() interface; see #4940.
+  return (int) sent;
 }
 
 /**
